final refactoring. changed files to obey style guide. made all components a .jsx
