+++ frontend/src/components/pages/CardDetails/Presentation.jsx
@@ -48,7 +48,7 @@ const useStyles = makeStyles((theme) => {
   };
 });
 
-const TopicProgressDetails = ({ topicProgress, reviews }) => {
+function TopicProgressDetails ({ topicProgress, reviews }) {
   return (
     <React.Fragment>
       {topicProgress.topicNeedsReview ? (
@@ -63,7 +63,7 @@ const TopicProgressDetails = ({ topicProgress, reviews }) => {
   );
 };
 
-const CardBasicDetails = ({ card }) => {
+function CardBasicDetails ({ card }) {
   const classes = useStyles();
 
   const dueTime = card.dueTime && new Date(card.dueTime).toLocaleString();
