+++ frontend/src/components/pages/CardDetails/ProjectDetails.jsx
