+++ frontend/src/components/pages/CardDetails/index.jsx
@@ -15,15 +15,6 @@ import {
 } from "../../../constants";
 
 function CardDetailsUnconnected({
-  // project,
-  // projectId,
-  // cardId,
-  // card,
-  // topicProgressId,
-  // topicProgress,
-  // topicReviews,
-  // projectReviews,
-  // handleClose,
   cards,
   projects,
   topicProgressArray,
@@ -156,59 +147,8 @@ function CardDetailsUnconnected({
 }
 
 const mapStateToProps = (state) => {
-  // const cardId = state.CardDetails.cardId;
-
-  // const card =
-  //   !!cardId & (state.apiEntities.cards !== undefined)
-  //     ? state.apiEntities.cards[cardId]
-  //     : null;
-
-  // const projectId =
-  //   card && card.contentTypeNice === "project" && card.recruitProject;
-  // const topicProgressId =
-  //   card && card.contentTypeNice === "topic" && card.topicProgress;
-
-  // const project =
-  //   !!projectId & (state.apiEntities.projects !== undefined)
-  //     ? state.apiEntities.projects[projectId]
-  //     : null;
-
-  // const topicProgress =
-  //   !!topicProgressId & (state.apiEntities.topicProgress !== undefined)
-  //     ? state.apiEntities.topicProgress[topicProgressId]
-  //     : null;
-
-  // const projectReviews =
-  //   (project !== null) &
-  //   (project !== undefined) &
-  //   (state.apiEntities.projectReviews !== undefined)
-  //     ? project.projectReviews
-  //         .map((reviewId) => {
-  //           return state.apiEntities.projectReviews[reviewId];
-  //         })
-  //         .filter((review) => review !== undefined)
-  //     : [];
-
-  // const topicReviews =
-  //   (topicProgress !== null) &
-  //   (topicProgress !== undefined) &
-  //   (state.apiEntities.topicReviews !== undefined)
-  //     ? topicProgress.topicReviews
-  //         .map((reviewId) => {
-  //           return state.apiEntities.topicReviews[reviewId];
-  //         })
-  //         .filter((review) => review !== undefined)
-  //     : [];
 
   return {
-    // cardId,
-    // card,
-    // topicProgressId,
-    // topicProgress,
-    // project,
-    // projectId,
-    // projectReviews,
-    // topicReviews,
     cards: state.apiEntities.cards,
     projects: state.apiEntities.projects,
     topicProgressArray: state.apiEntities.topicProgress,
